+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 - biff.xtdb: add schema-module
 - biff.core: accept maps for :biff.core/init values
 - biff.tasks: format task adds blank lines to separate split form pairs
-- biff.tasks: add :biff/impl-visibility clj-kondo rule
+- biff.tasks: add `:biff/impl-visibility` clj-kondo rule
+- biff.tasks: init task initializes git repo
 
 **Breaking** [2.0.0-rc23](RELEASES/2.0.0-rc23.md) (2026-08-29):